+++ staticfiles/conf.py
@@ -1,3 +1,0 @@
-from staticfiles.settings import StaticfilesSettings
-
-settings = StaticfilesSettings("STATICFILES")