+++ staticfiles/finders.py
@@ -1,4 +1,5 @@
 import os
+from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.core.files.storage import Storage
 from django.utils.datastructures import SortedDict
@@ -7,7 +8,6 @@ from django.utils.importlib import import_module
 from django.utils._os import safe_join
 
 from staticfiles import utils, storage
-from staticfiles.conf import settings
 
 _finders = SortedDict()
 
@@ -27,7 +27,7 @@ class BaseFinder(object):
         """
         raise NotImplementedError()
 
-    def list(self, ignore_patterns=[]):
+    def list(self, ignore_patterns):
         """
         Given an optional list of paths to ignore, this should return
         a three item iterable with path, prefix and a storage instance.
@@ -132,7 +132,7 @@ class AppDirectoriesFinder(BaseFinder):
         """
         List all files in all app storages.
         """
-        for storage in self.storages.values():
+        for storage in self.storages.itervalues():
             if storage.exists(''): # check if storage location exists
                 for path in utils.get_files(storage, ignore_patterns):
                     yield path, storage