+++ staticfiles/urls.py
@@ -5,6 +5,7 @@ from django.core.exceptions import ImproperlyConfigured
 
 urlpatterns = []
 
+
 def static(prefix, view='django.views.static.serve', **kwargs):
     """
     Helper function to return a URL pattern for serving files in debug mode.
@@ -17,14 +18,13 @@ def static(prefix, view='django.views.static.serve', **kwargs):
     ) + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
 
     """
-    if not settings.DEBUG:
+    # No-op if not in debug mode or an non-local prefix
+    if not settings.DEBUG or (prefix and '://' in prefix):
         return []
     elif not prefix:
         raise ImproperlyConfigured("Empty static prefix not permitted")
-    elif '://' in prefix:
-        raise ImproperlyConfigured("URL '%s' not allowed as static prefix" % prefix)
     return patterns('',
-        url(r'^%s(?P<path>.*)$' % re.escape(prefix.lstrip('/')), view, **kwargs),
+        url(r'^%s(?P<path>.*)$' % re.escape(prefix.lstrip('/')), view, kwargs=kwargs),
     )
 
 def staticfiles_urlpatterns(prefix=None):